fixing warnings on 64bit compiler (just casting size_t to int)
git-svn-id: https://svn.code.sf.net/p/personalrobots/code/pkg/trunk@23037 f5854215-dd47-0410-b2c4-cdd35faa7885

Former-commit-id: 752c134ed06bb2ab28a26527b751673aff567f46

+++ stacks/robot_model/urdf/include/urdf/color.h
@@ -77,7 +77,7 @@ public:
 
     if (rgba.size() != 4)
     {
-      ROS_ERROR("Color contains %i elements instead of 4 elements", rgba.size());
+	ROS_ERROR("Color contains %i elements instead of 4 elements", (int)rgba.size());
       return false;
     }
